Moved SolrNet integration tests to separate project.
Moved integration tests for each IOC component into their own classes.
Consistent naming of test fixtures.
