+++ AutofacContrib.SolrNet.Tests/AutofacIntegrationFixture.cs
@@ -1,0 +1,56 @@
+﻿// 
+
+using System;
+using System.Collections.Generic;
+using Autofac;
+using MbUnit.Framework;
+using SolrNet;
+
+namespace AutofacContrib.SolrNet.Tests {
+    [TestFixture]
+    [Category("Integration")]
+    public class AutofacIntegrationFixture {
+        [Test]      
+        public void Ping_And_Query()
+        {
+            var builder = new ContainerBuilder();
+            builder.RegisterModule(new SolrNetModule("http://localhost:8983/solr"));
+            var container = builder.Build();
+            var solr = container.Resolve<ISolrOperations<AutofacFixture.Entity>>();
+            solr.Ping();
+            Console.WriteLine(solr.Query(SolrQuery.All).Count);
+        }
+
+        [Test]
+        public void DictionaryDocument()
+        {
+            var builder = new ContainerBuilder();
+            builder.RegisterModule(new SolrNetModule("http://localhost:8983/solr"));
+            var container = builder.Build();
+            var solr = container.Resolve<ISolrOperations<Dictionary<string, object>>>();
+            var results = solr.Query(SolrQuery.All);
+            Assert.GreaterThan(results.Count, 0);
+            foreach (var d in results)
+            {
+                Assert.GreaterThan(d.Count, 0);
+                foreach (var kv in d)
+                    Console.WriteLine("{0}: {1}", kv.Key, kv.Value);
+            }
+        }
+
+        [Test]
+        public void DictionaryDocument_add()
+        {
+            var builder = new ContainerBuilder();
+            builder.RegisterModule(new SolrNetModule("http://localhost:8983/solr"));
+            var container = builder.Build();
+            var solr = container.Resolve<ISolrOperations<Dictionary<string, object>>>();
+            solr.Add(new Dictionary<string, object> {
+                {"id", "ababa"},
+                {"manu", "who knows"},
+                {"popularity", 55},
+                {"timestamp", DateTime.UtcNow},
+            });
+        }
+    }
+}
