+++ Castle.Facilities.SolrNetIntegration.Tests/Castle.Facilities.SolrNetIntegration.Tests.csproj
@@ -80,7 +80,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Tests.cs" />
+    <Compile Include="CastleIntegrationFixture.cs" />
+    <Compile Include="CastleFixture.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
